app/requirement.txt
psycopg2-binary==2.9.9
faker==24.2.0
tqdm==4.66.2